Modify PromptPart not to use Chip compnent
Implement PromptPart with div and css instead of Material UI's
Chip component, because Chip has the animation to gradually change the background
color. This animation runs on every rendering Chip even when dropped it on Droppable.
Eventually drag-and-dropping Chip comes annoying UI.
To avoid it, modify PromptPart component not to use Chip.